[#109]Rename(project): merge 후 project 이름 및 경로 재변경, 코드 포매팅
